requirements.txt
python-docx
reportlab